+++ app/components/views/GetStartedPage/SetupWallet/SetMixedAcctPage/SetMixedAcctPage.jsx
@@ -6,6 +6,7 @@ import { KeyBlueButton } from "buttons";
 import { MIXED_ACCOUNT, CHANGE_ACCOUNT } from "constants";
 import { GoBackMsg } from "../../messages";
 import { useDaemonStartup, useMountEffect, useAccounts } from "hooks";
+// change this import and move all styles to component's css module
 import GetStartedStyles from "../../GetStarted.module.css";
 import styles from "./SetMixedAcctPage.module.css";
