--- app/components/views/GetStartedPage/SetupWallet/SetupWalletPage.jsx
@@ -1,13 +1,0 @@
-import { classNames } from "pi-ui";
-import styles from "../GetStarted.module.css";
-
-const CreateWalletPage = ({ StateComponent }) => (
-  <div className={classNames(styles.getstarted, styles.content)}>
-    {StateComponent &&
-      (React.isValidElement(StateComponent) ? StateComponent : (
-          <StateComponent />
-        ))}
-  </div>
-);
-
-export default CreateWalletPage;
